minor improvements fetch_data.py

+++ scripts/fetch_data.py
@@ -15,6 +15,7 @@ import pandas as pd
 import numpy as np
 from dataclasses import dataclass
 from datetime import datetime
+import argparse
 
 if os.environ.get('CDF_LIB', '') == '':
     print('No CDF_LIB environment variable found for CDF file processing.')
@@ -22,9 +23,9 @@ if os.environ.get('CDF_LIB', '') == '':
 from spacepy import pycdf
 
 
-DATA_RAW_DIR = '../data_raw'
-DATA_PROCESSED_DIR = '../data_processed'
-OUTPUT_FILE_WITHOUT_EXT = 'preprocessed'
+DATA_RAW_DIR = pathlib.Path('../data_raw')
+DATA_PROCESSED_DIR = pathlib.Path('../data_processed')
+OUTPUT_FILE_WITHOUT_EXT = pathlib.Path('preprocessed')
 
 
 @dataclass
@@ -35,17 +36,19 @@ class RawCDF:
     data: pycdf.CDF
 
     def count_events(self) -> int:
-        total_channels = 31 + 9 + 9  # TODO: Rewrite in terms of cdf Vars
+        # TODO: Rewrite in terms of cdf Vars
+        total_channels = 31 + 9 + 9
         return total_channels * len(self.data["TIME_UTC"])
 
     @staticmethod
-    def parse_date(filename: str) -> datetime:
+    def parse_date(filename: pathlib.Path) -> datetime.date:
         date_string = filename[-12:-4]
-        format = '%Y%m%d'
-        return datetime.strptime(date_string, format).date()
+        format_ = '%Y%m%d'
+        dt = datetime.strptime(date_string, format_).date()
+        return dt
 
     @staticmethod
-    def parse_type(filename: str) -> str:
+    def parse_type(filename: pathlib.Path) -> str:
         # FixMe: Non exhaustive match
         return 'science' if filename[8:10] == 'sc' else 'housekeeping'
 
@@ -88,7 +91,7 @@ class RawCDF:
         return df
 
 
-def make_dir(dir: str) -> bool:
+def make_dir(dir: pathlib.Path) -> bool:
     if 0 != subprocess.call(f"mkdir -p {dir}",
                             shell=True):
         print(f"Error creating {dir} directory", file=sys.stderr)
@@ -96,7 +99,7 @@ def make_dir(dir: str) -> bool:
     return True
 
 
-def remove_dir(dir: str) -> bool:
+def remove_dir(dir: pathlib.Path) -> bool:
     if 0 != subprocess.call(f"rm -rd {dir}",
                             shell=True):
         print(f"Error removing {dir} directory", file=sys.stderr)
@@ -104,7 +107,7 @@ def remove_dir(dir: str) -> bool:
     return True
 
 
-def remove_file(file: str) -> bool:
+def remove_file(file: pathlib.Path) -> bool:
     if 0 != subprocess.call(f"rm -f {file}",
                             shell=True):
         print(f"Error removing {file} file", file=sys.stderr)
@@ -112,7 +115,7 @@ def remove_file(file: str) -> bool:
     return True
 
 
-def fetch_data(target_dir: str) -> bool:
+def fetch_data(target_dir: pathlib.Path) -> bool:
     if 0 != subprocess.call(f"wget -r --timestamping --continue --user=ifjagh --ask-password ftp://ftptrans.psi.ch/to_radem/ -nd -np -P {target_dir}",
                             shell=True):
         print("Error fetching data", file=sys.stderr)
@@ -120,7 +123,7 @@ def fetch_data(target_dir: str) -> bool:
     return True
 
 
-def extract_data(source_dir: str, target_dir: str) -> bool:
+def extract_data(source_dir: pathlib.Path, target_dir: pathlib.Path) -> bool:
     if 0 != subprocess.call(f"for f in {source_dir}/*.tar.gz; do tar -xvf \"$f\" -C {target_dir}; done;",
                             shell=True):
         print("Error extracting tar files", file=sys.stderr)
@@ -128,7 +131,7 @@ def extract_data(source_dir: str, target_dir: str) -> bool:
     return True
 
 
-def process_data(source_dir: str) -> pd.DataFrame:
+def process_data(source_dir: pathlib.Path) -> pd.DataFrame:
     cdfs = [
         RawCDF(name=path.name,
                date=RawCDF.parse_date(path.name),
@@ -138,12 +141,16 @@ def process_data(source_dir: str) -> pd.DataFrame:
     ]
 
     science_cdfs = [cdf for cdf in cdfs if cdf.type_ == 'science']
+    science_cdfs.sort(key=lambda cdf: cdf.date)
+    print(science_cdfs)
 
     print(f"{len(science_cdfs)} CDFs found. Get a coffee because this may take a while...")
 
-    df = pd.concat([cdf.to_dataframe(i)
-                    for i, cdf in enumerate(science_cdfs)])
+    A = [cdf.to_dataframe(i)
+         for i, cdf in enumerate(science_cdfs)]
 
+    df = pd.concat(A)
+    df.sort_values(by=['time', 'event_type', 'channel'], inplace=True)
     return df
 
 
@@ -151,17 +158,31 @@ def filter_data(df: pd.DataFrame) -> None:
     df.query("time >= '2023-09-01'", inplace=True)
 
 
-def save_csv_data(df: pd.DataFrame, target_dir: str, file_without_ext: str):
+def save_csv_data(df: pd.DataFrame, target_dir: pathlib.Path, file_without_ext: str):
     df.to_csv(f'{target_dir}/{file_without_ext}.csv', index=False)
 
 
-def save_hdf_data(df: pd.DataFrame, target_dir: str, file_without_ext: str):
+def save_hdf_data(df: pd.DataFrame, target_dir: pathlib.Path, file_without_ext: str):
     df.to_hdf(f'{target_dir}/{file_without_ext}.h5',
               key='time', format="table")
 
 
-def pipeline():
+def args_parser() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description='Fetch data from PSI FTP server and save it to HDF5 and CSV files.')
+    parser.add_argument('--skip-fetch', action='store_true',
+                        help='Skip fetching data from PSI FTP server')
+    parser.add_argument('--skip-cleanup', action='store_true',
+                        help='Skip cleanup of temporary files')
+    return parser.parse_args()
+
+
+def pipeline(args: argparse.Namespace | None = None):
+    if args is None:
+        args = args_parser()
+
     try:
+        # "../data_tmp_1711394609"
         temp_dir = pathlib.Path(f'../data_tmp_{int(time.time())}')
         if not make_dir(temp_dir):
             return
@@ -169,8 +190,9 @@ def pipeline():
         if not make_dir(DATA_RAW_DIR):
             return
 
-        if not fetch_data(DATA_RAW_DIR):
-            return
+        if not args.skip_fetch:
+            if not fetch_data(DATA_RAW_DIR):
+                return
 
         if not extract_data(DATA_RAW_DIR, temp_dir):
             return
@@ -184,16 +206,18 @@ def pipeline():
 
         save_csv_data(df, DATA_PROCESSED_DIR, OUTPUT_FILE_WITHOUT_EXT)
         save_hdf_data(df, DATA_PROCESSED_DIR, OUTPUT_FILE_WITHOUT_EXT)
+
+        print("SUCCESS")
     except KeyboardInterrupt:
         print("Interrupted by user")
         remove_file(f'{DATA_PROCESSED_DIR}/{OUTPUT_FILE_WITHOUT_EXT}.csv')
         remove_file(f'{DATA_PROCESSED_DIR}/{OUTPUT_FILE_WITHOUT_EXT}.h5')
     finally:
-        if not remove_dir(temp_dir):
-            return
-
-    print("SUCCESS")
+        if not args.skip_cleanup:
+            if not remove_dir(temp_dir):
+                return
 
 
 if __name__ == "__main__":
-    pipeline()
+    args = args_parser()
+    pipeline(args)
